Update filter and feeds layout
